Improved checking of valid vs. invalid People values, more informative warnings.

+++ sfm/ui/models.py
@@ -2,6 +2,7 @@ import datetime
 import gzip
 import json
 import re
+import os
 import time
 
 import requests
@@ -218,8 +219,8 @@ class TwitterUserItem(m.Model):
                    'screen_name', 'followers_count', 'friends_count',
                    'retweet_count', 'hashtags', 'in_reply_to_screen_name',
                    'mentions', 'twitter_url', 'is_retweet_strict',
-                   'is_retweet', 'coordinates', 'text', 'url1', 'url1_expanded', 'url2',
-                   'url2_expanded']
+                   'is_retweet', 'coordinates', 'text',
+                   'url1', 'url1_expanded', 'url2', 'url2_expanded']
 
     def __unicode__(self):
         return '<useritem (%s)>' % (self.id)
@@ -394,6 +395,7 @@ documentation</a> for more information.""")
                                      Please select a different user or mark
                                      this filter as inactive.''' %
                                   (conflicting_tfs[0].id, self.user.username))
+
         # update line-breaks in words with spaces
         if self.words != '':
             wrd = self.words
@@ -431,22 +433,30 @@ documentation</a> for more information.""")
                 raise ValidationError("Invalid bounding box; each \
                         long/lat value must be between -180 and 180")
 
+
 @receiver(post_save, sender=TwitterFilter)
 def call_create_conf(sender, instance, **kwargs):
-    # Removing the process so that supervisor may implement the updates in
-    # twitter filter dynamically as previously the updates were ignored by
-    # supervisor in case an active filter was modified
-    ui.utils.remove_process_group(instance.id)
-    if instance.is_active is True:
-        create_conf_file(instance.id)
-        time.sleep(1)
-        ui.utils.reload_config()
-        ui.utils.add_process_group(instance.id)
-    else:
-        delete_conf_file(instance.id)
+    # Remove and (in the case of an active filter) re-add
+    # the supervisor process group.  Otherwise the configuration
+    # file updates are ignored by supervisor.
+    #
+    # TODO: Consider not just swallowing the condition
+    # where Supervisor isn't running
+    if os.path.exists(settings.SUPERVISOR_UNIX_SOCKET_FILE):
+        ui.utils.remove_process_group(instance.id)
+        if instance.is_active is True:
+            create_conf_file(instance.id)
+            time.sleep(1)
+            ui.utils.reload_config()
+            ui.utils.add_process_group(instance.id)
+        else:
+            delete_conf_file(instance.id)
 
 
 @receiver(post_delete, sender=TwitterFilter)
 def call_delete_conf(sender, instance, **kwargs):
-    delete_conf_file(instance.id)
-    ui.utils.remove_process_group(instance.id)
+    # It might be okay in the case of deleting,
+    # to swallow the condition where Supervisor isn't running.
+    if os.path.exists(settings.SUPERVISOR_UNIX_SOCKET_FILE):
+        delete_conf_file(instance.id)
+        ui.utils.remove_process_group(instance.id)
